requirements.txt
numpy
opencv-python
PyWavelets
pycryptodome
streamlit
pandas
